use double in intermediate calculations

+++ src/main/java/suite/trade/MovingAverage.java
@@ -8,7 +8,7 @@ public class MovingAverage {
 	private Matrix mtx = new Matrix();
 
 	// moving average convergence/divergence
-	public float[] macd(float[] prices, float alpha0, float alpha1) {
+	public float[] macd(float[] prices, double alpha0, double alpha1) {
 		float[] emas0 = exponentialMovingAvg(prices, alpha0); // long-term
 		float[] emas1 = exponentialMovingAvg(prices, alpha1); // short-term
 		return mtx.sub(emas1, emas0);
@@ -26,11 +26,11 @@ public class MovingAverage {
 		return To.arrayOfFloats(movingAvgs, lma -> (float) Math.exp(lma));
 	}
 
-	public float[] exponentialMovingAvg(float[] prices, float alpha) {
+	public float[] exponentialMovingAvg(float[] prices, double alpha) {
 		float[] emas = new float[prices.length];
-		float ema = prices[0];
+		double ema = prices[0];
 		for (int day = 0; day < prices.length; day++)
-			emas[day] = ema += alpha * (prices[day] - ema);
+			emas[day] = (float) (ema += alpha * (prices[day] - ema));
 		return emas;
 	}
